+++ controllers/courses.js
@@ -49,13 +49,22 @@ exports.getCourse = asyncHandler(async (req,res,next)=>{
 //@route POST /api/v1/bootcamp/:bootcampId/courses
 //@access Private
 exports.addCourse = asyncHandler(async (req,res,next)=>{
- req.body.bootcamp = req.params.bootcampId
+ req.body.bootcamp = req.params.bootcampId;
+ req.body.user = req.user.id;
 
     const bootcamp =await Bootcamp.findById(req.params.bootcampId).populate()
     if(!bootcamp){
      return next (
          new ErrorResponse(`No bootcamp with id of ${req.params.bootcampId}`,404));
     }
+    //make sure logged in user is the owner of the bootcamp
+if(bootcamp.user.toString() != req.user.id && req.user.role != 'admin'){
+    return next (
+        new ErrorResponse(`User with id of ${req.user.id} not an owner of this bootcamp`,401)
+    )
+}
+
+
     const course = await Course.create(req.body);
  
      res.status(200).json({
@@ -76,6 +85,14 @@ exports.updateCourse = asyncHandler(async (req,res,next)=>{
         return next (
             new ErrorResponse(`No course with id of ${req.params.bootcampId}`,404));
        }
+
+       //make sure logged in user is the owner of the bootcamp
+if(course.user.toString() != req.user.id && req.user.role != 'admin'){
+    return next (
+        new ErrorResponse(`User with id of ${req.user.id} not an owner of this bootcamp`,401)
+    )
+}
+
 
       course = await Course.findByIdAndUpdate(req.params.id,req.body,{
         new :true,
@@ -100,6 +117,13 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
         new ErrorResponse(`No course with the id of ${req.params.id}`, 404)
       );
     }
+          //make sure logged in user is the owner of the bootcamp
+if(course.user.toString() != req.user.id && req.user.role != 'admin'){
+    return next (
+        new ErrorResponse(`User with id of ${req.user.id} not an owner of this bootcamp`,401)
+    )
+}
+
   
     
     await course.deleteOne();
